seg faults

+++ src/extlibs/enlsvg/src/Fraction.cpp
@@ -22,35 +22,35 @@ Fraction operator-(const Fraction& o, const Fraction& o2) {
     return Fraction(o.n*o2.d - o2.n*o.d, o.d*o2.d);
 }
     
-Fraction operator*(const Fraction& o, int o2) {
+Fraction operator*(const Fraction& o, long o2) {
     return Fraction(o.n*o2, o.d);
 }
 
-Fraction operator/(const Fraction& o, int o2) {
+Fraction operator/(const Fraction& o, long o2) {
     return Fraction(o.n, o.d*o2);
 }
 
-Fraction operator+(const Fraction& o, int o2) {
+Fraction operator+(const Fraction& o, long o2) {
     return Fraction(o.n + o2*o.d, o.d);
 }
 
-Fraction operator-(const Fraction& o, int o2) {
+Fraction operator-(const Fraction& o, long o2) {
     return Fraction(o.n - o2*o.d, o.d);
 }
 
-Fraction operator*(int o, const Fraction& o2) {
+Fraction operator*(long o, const Fraction& o2) {
     return Fraction(o*o2.n, o2.d);
 }
 
-Fraction operator/(int o, const Fraction& o2) {
+Fraction operator/(long o, const Fraction& o2) {
     return Fraction(o*o2.d, o2.n);
 }
 
-Fraction operator+(int o, const Fraction& o2) {
+Fraction operator+(long o, const Fraction& o2) {
     return Fraction(o*o2.d + o2.n, o2.d);
 }
 
-Fraction operator-(int o, const Fraction& o2) {
+Fraction operator-(long o, const Fraction& o2) {
     return Fraction(o*o2.d - o2.n, o2.d);
 }
 
@@ -60,14 +60,14 @@ ostream& operator<< (ostream& stream, Fraction& obj) {
 
 
 Fraction parseFraction(string s) {
-    int i=s.find_first_of('/');
+    long i=s.find_first_of('/');
     if (i == string::npos) {
-        int n;
+        long n;
         stringstream(s) >> n;
 
         return Fraction(n);
     } else {
-        int n, d;
+        long n, d;
         stringstream(s.substr(0,i)) >> n;
         stringstream(s.substr(i+1)) >> d;
